Fixed: “Books I own in the series” in a very specific scenario + some cleanup of bookdetails
I broke book details on purpose. I’ve been trying to figure out how to best manage the summary that can sometimes be like 1 secntence and sometimes seemingly endless.

I’m still convinced it’s best to keep it constrained to the height of the information panel on the left, but I think I’ll go back to the original idea, which was to add a read more button.