0.1.5.1
Way too many things changed here for this to be a 0.0.0.1 update, but this is kind of a broken commit so I hesitate to give it a real number. If few more things was slightly more polished + the spreadsheet wasn’t such a performance hog, this commit might’ve been the first 1.0 version.

I forgot to make commits for a long time and now I’m wanting to get rid of Handsontable for it’s slow scroll performance so I’m making this comment to explore other options.