Fixed #23: wishlist scraping errored out due to fetching second lvl domain names like “.co.uk” wrong
The extension only fetched the last part so “.co.uk” turned into “.uk” and all of a sudden it didn’t know how to process the dates.

This error could happen during store page extraction, but mainly in the wishlist where it tries to “fix” the dates for every single book.