Fixed #28: Series/sub page fails to show the right content
The issue I believe comes from the fact that the series page is created using the book.series prop and series/subpage generates its content from library.series and if library series doesn’t have the series listed, it’s going to throw an error and also fail to find any books.

Why this discrepancy might happen is escaping me right now… My best guess is that the whole series page doesn’t exist anymore?

The fix is to make sure that the series page only shows books that are in the library.series array…