Cleaning wishlist page filtering and sorting
Basically just removing unnecessary filters and sorters from the wishlist.

A small thing, but probably the biggest change is that you can now filter plus catalog books in your wishlist. Of course, whether or not they are still in the plus catalog depends on how up to date your extracted data is.