Image editor: added some new animations + fixed an image rendering issue
I almost can’t believe this rendering issue that I just fixed… I had already given up on it multiple times because there just didn’t seem to be a fix…

The issue was that during not all but most css animations the images would look reaaaally sharp for the duration of the animation and then return to normal.
It really looked a lot like it had image-rendering: pixelated during the animation…

I’ve been working on this in Chrome and so I have no idea if the issue was there in other browsers.

The fix ended up being background color on the animated element… Like literally just added “background-color” to the img elements and problem soved…