Gallery chunk name override (#65) + minor fixed
1. There should be less need to replace chunk save-gallery.vue → var list — with new chunk names. 

Still a bit of a house of cards that may or may not cause the same thing later if imports change.

The "thing" being a need to build > get dist/chunks to the list variable in save-gallery.vue > build... 

Which if necessary, is so only when testing standalone gallery saving or the standalone gallery itself... or of course when doing a final release.

And you should never save the standalone gallery while serving. Build is way more optimized.

2. Removed some more .[ typos

3. Disabled trailing whitespace for editors that use .vscode/settings.js or .editorconfig

4. Fixed a small issue with the search options list where you couldn't click directly directly below it to close it... Also added some minor adjustments to the fixed overlay height when the options list is open.

5. Some code cleanup (deletion)... Mostly removing commented out code.

6. Fixed gallery active sorter highlight issue when searching...