Fixed "my books in the series" book links
The issue was that when you clicked the links, the book details stayed on the book that was opened before you clicked the link... rather than showing the book that was clicked.

> This issue was introduced at some point when I redid the links when I started working on v.0.2.7

I also fixed a small issue with the series title above those links so that they work with the same kind of logic.