Added a console to the extraction settings
There's an icon in the bottom right that opens the console. 

One of the perks of this is that it seems to filter out most if not all errors generated by audible.com.

Also it will be easier to tell users to click this than to tell them to open the browser's console which is slightly different in different browsers.