Improved browser context menu
# Before
Items 1 and 2 were missing if the user had not yet successfully extracted data. Or if the data was deleted.

# After
Item 1 leads to audible.com/library. Item 2 is grayed out with a small message saying it's available after extraction.

So now you kinda now what to expect if you happen to open the context menu before extracting data.

# Also...
The new user message about the context menu now has an image to make it clearer...