Added book detail settings
It started with the idea that the (long) titles are kinda unwieldy, so why I don't I add the option to use the shorter title. 

At first I attempted to put the setting inside that almost useless book details info icon, but it had some issues so I decided to overlay it on top of book details instead and made that icon into a button that opens the settings. 

And mostly because there was room for it, I then added a bunch of settings for hiding elements... Some of it doesn't necessary make a whole lot of sense, but whatever...