Series extraction bug fix: sometimes book number ended up being extracted as if it's a series
The previous system harnessed the power of luck with the assumption that first element is always the title and the next one its number, but that isn't always the case, sometimes it can be "title, title number" or it could even be "title, title".

Now we detect titles by fetching its href first. If it's a link, it's the title, otherwise it's the number if regex match says it has numbers in it.

Since the number element always comes after the title, the number is then pushed to the previously extracted series, which is constructed when the title is found.